Opening home page under Linux and OSX.-
