+++ pwdhash/console.py
@@ -11,6 +11,8 @@ def go (pwd_gen):
 
     pwd_gen     the PwdHash generator instance to use.-
     """
+    from pwdhash.platform import copy_to_clipboard
+
     domain    = raw_input ("For domain: ").strip ( )
     generated = pwd_gen.generate (domain)
 
@@ -23,43 +25,3 @@ def go (pwd_gen):
         print "Password copied to clipboard."
     else:
         print generated
-
-
-
-def copy_to_clipboard (string):
-    """
-    Returns 'True' if 'string' was correctly copied to the system's clipboard.-
-    """
-    #
-    # an external program is used for copying the password to the clipboard
-    #
-    is_copied = False
-
-    if sys.platform == "darwin":
-        #
-        # on OSX
-        #
-        clip_copy_exe = "pbcopy"
-    elif 'DISPLAY' in os.environ:
-        #
-        # on Linux/Un*x
-        #
-        clip_copy_exe = "xclip"
-
-    try:
-        pb = subprocess.Popen(clip_copy_exe,
-                              stdin=subprocess.PIPE,
-                              stdout=open("/dev/null", "w"),
-                              stderr=open("/dev/null", "w"))
-        pb.communicate (string)
-        pb.wait ( )
-        if pb.returncode == 0:
-            is_copied = True
-        else:
-            is_copied = False
-            logging.warning ("Install '%s' for clipboard support" % clip_copy_exe)
-    except:
-        is_copied = False
-
-    return is_copied
-
