+++ requirements.txt
@@ -2,6 +2,6 @@ CherryPy==3.6.0
 Jinja2==2.7.3
 MarkupSafe==0.23
 SQLObject==2.1.0
-apsw==3.8.5-r1
+apsw
 nose==1.3.4
 wsgiref==0.1.2
